src/usr/lib/libre-workspace/portal/requirements.txt
django
python-ldap
django-auth-ldap
gunicorn
requests
django-oidc-provider
image
django_otp
qrcode
Jinja2
djangorestframework
drf-spectacular
# For not be dependend on cdn:
drf-spectacular-sidecar 
